[GREEN] Implement field and authorization validation in Listing service

+++ src/main/java/id/ac/ui/cs/advprog/backendbuysell/service/ListingServiceImpl.java
@@ -3,14 +3,17 @@ package id.ac.ui.cs.advprog.backendbuysell.service;
 import id.ac.ui.cs.advprog.backendbuysell.dto.ListingSearchRequestDTO;
 import id.ac.ui.cs.advprog.backendbuysell.dto.ListingSearchResponseDTO;
 import id.ac.ui.cs.advprog.backendbuysell.enums.ListingStatus;
+import id.ac.ui.cs.advprog.backendbuysell.exception.FieldValidationException;
+import id.ac.ui.cs.advprog.backendbuysell.exception.ForbiddenException;
 import id.ac.ui.cs.advprog.backendbuysell.model.Listing;
 import id.ac.ui.cs.advprog.backendbuysell.repository.ListingRepository;
 import id.ac.ui.cs.advprog.backendbuysell.utils.ListingSearchQueryBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.Errors;
+import org.springframework.validation.Errors;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -23,7 +26,12 @@ public class ListingServiceImpl implements ListingService {
     @Autowired
     ListingRepository listingRepository;
 
-    public Listing create(Listing listing) {
+    public Listing create(Listing listing, String sellerId) {
+        listing.setSellerId(sellerId);
+        Errors validationResult = listing.validate();
+        if (validationResult.hasErrors()) {
+            throw new FieldValidationException(validationResult.getAllErrors(), "Validation Error");
+        }
         return listingRepository.save(listing);
     }
 
@@ -43,16 +51,32 @@ public class ListingServiceImpl implements ListingService {
         return listingRepository.findById(id);
     }
 
-    public Listing update(Long id, Listing updatedListing) {
-        if (!listingRepository.existsById(id)) {
+    public Listing update(Long id, Listing updatedListing, String sellerId) {
+        Optional<Listing> oldResult = listingRepository.findById(id);
+        if (oldResult.isEmpty()) {
             throw new NoSuchElementException();
+        }
+        if (!isAuthenticated(oldResult.get(), sellerId)) {
+            throw new ForbiddenException("User is not authorized to perform action on this listing");
+        }
+        updatedListing.setSellerId(sellerId);
+        Errors validationResult = updatedListing.validate();
+        if (validationResult.hasErrors()) {
+            throw new FieldValidationException(validationResult.getAllErrors(), "Validation Error");
         }
         updatedListing.setId(id);
         return listingRepository.save(updatedListing);
     }
 
-    public Listing delete(Long id) {
+    public Listing delete(Long id, String sellerId) {
         Listing listing = listingRepository.findById(id).orElseThrow();
+        if (!isAuthenticated(listing, sellerId)) {
+            throw new ForbiddenException("User is not authorized to perform action on this listing");
+        }
+        Errors validationResult = listing.validate();
+        if(validationResult.hasErrors()){
+            throw new FieldValidationException(validationResult.getAllErrors(), "Validation Error");
+        }
         listingRepository.deleteById(id);
         return listing;
     }
@@ -69,9 +93,20 @@ public class ListingServiceImpl implements ListingService {
         return this.getAll(requestDTO);
     }
 
-    public void setStatus(Long id, String status) {
+    public void setStatus(Long id, String status, String sellerId) {
         Listing listing = listingRepository.findById(id).orElseThrow();
+        if (!isAuthenticated(listing, sellerId)) {
+            throw new ForbiddenException("User is not authorized to perform action on this listing");
+        }
         listing.setStatus(status);
+        Errors validationResult = listing.validate();
+        if(validationResult.hasErrors()){
+            throw new FieldValidationException(validationResult.getAllErrors(), "Validation Error");
+        }
         listingRepository.save(listing);
+    }
+
+    private boolean isAuthenticated(Listing listing, String sellerId) {
+        return listing.getSellerId().equals(sellerId);
     }
 }
